Cleaning and moving props part 7
