+++ src/main/java/com/paneedah/mwc/models/props/desks/DeskShelfAlt1.java
@@ -1,4 +1,4 @@
-package com.paneedah.mwc.models.props;
+package com.paneedah.mwc.models.props.desks;
 
 import com.paneedah.mwc.models.IModernModel;
 import net.minecraft.client.model.ModelBase;
