+++ src/main/java/com/paneedah/mwc/models/props/desks/WhiteDeskLeft.java
@@ -1,4 +1,4 @@
-package com.paneedah.mwc.models.props;
+package com.paneedah.mwc.models.props.desks;
 
 import com.paneedah.mwc.models.IModernModel;
 import net.minecraft.client.model.ModelBase;
@@ -12,19 +12,19 @@ import net.minecraft.entity.Entity;
  * @author Developer: Desoroxx
  */
 @SuppressWarnings({"FieldCanBeLocal", "NullableProblems"})
-public final class WhiteDeskRight extends ModelBase implements IModernModel {
+public final class WhiteDeskLeft extends ModelBase implements IModernModel {
 
     private final ModelRenderer bone;
 
-    public WhiteDeskRight() {
+    public WhiteDeskLeft() {
         textureWidth = 128;
         textureHeight = 128;
 
         bone = new ModelRenderer(this);
         bone.setRotationPoint(0.0F, 24.0F, 0.0F);
+        bone.cubeList.add(new ModelBox(bone, 0, 0, -8.1F, -15.9F, 6.1F, 2, 16, 2, 0.0F, true));
         bone.cubeList.add(new ModelBox(bone, 0, 0, -8.0F, -16.0F, -5.0F, 16, 2, 13, 0.0F, true));
-        bone.cubeList.add(new ModelBox(bone, 0, 0, 6.1F, -15.9F, -5.1F, 2, 16, 2, 0.0F, true));
-        bone.cubeList.add(new ModelBox(bone, 0, 0, 6.1F, -15.9F, 6.1F, 2, 16, 2, 0.0F, true));
+        bone.cubeList.add(new ModelBox(bone, 0, 0, -8.1F, -15.9F, -5.1F, 2, 16, 2, 0.0F, true));
     }
 
     @Override
